Add files via upload
A snowman is drawn using 3 adjacent and non-overlapping circles of diminishing radius, with colinear 10 centers. In this question you need to fill an image with random, non-overlapping snowpeople of varying sizes, using multiple threads.
The classic board game “Snakes and Ladders” (or “Chutes and Ladders”) is based on a 10x10 grid. A 10 player starts in the lower-left grid cell, and based on a random roll (in the range 1–6) advances that many grid cells horizontally, potentially including moving to the cell in the next row upward if they would go off the board horizontally. l simulate concurrent game play and board modification. First, build a board consisting of a 10 × 10 array of CELL objects. Initialize the board with 10 snakes and 9 ladders, randomly created, with the condition that snake heads and tails and ladder tops and bottoms must always refer to different cells and never to the starting or ending cells, and snakes (and ladders) must always go between different rows. Code should accept parameters k, j, and s, where s is the number of seconds to let the simulation run. After s seconds have passed all threads should stop execution. Sort the thread logs by timestamp and emit an ordered list of operations, as per the following (note that cells are uniquely identified by their y ∗ 10 + x coordinate, 0-based of course, and the comments are just here to describe the format). Code should accept parameters k, j, and s, where s is the number of seconds to let the simulation run.